@@ -1,9 +1,25 @@
 module github.com/anoideaopen/foundation/test/integration
 
-go 1.23.6
+go 1.24.0
+
+tool (
+	github.com/IBM/idemix/tools/idemixgen
+	github.com/anoideaopen/acl
+	github.com/anoideaopen/channel-transfer
+	github.com/anoideaopen/foundation/test/chaincode/cc
+	github.com/anoideaopen/foundation/test/chaincode/fiat
+	github.com/anoideaopen/foundation/test/chaincode/industrial
+	github.com/anoideaopen/robot
+	github.com/hyperledger/fabric/cmd/configtxgen
+	github.com/hyperledger/fabric/cmd/cryptogen
+	github.com/hyperledger/fabric/cmd/discover
+	github.com/hyperledger/fabric/cmd/orderer
+	github.com/hyperledger/fabric/cmd/osnadmin
+	github.com/hyperledger/fabric/cmd/peer
+	github.com/hyperledger/fabric/integration/chaincode/simple/cmd
+)
 
 require (
-	github.com/IBM/idemix v0.0.2-0.20240913182345-72941a5f41cd
 	github.com/anoideaopen/acl v0.1.0
 	github.com/anoideaopen/channel-transfer v0.1.0
 	github.com/anoideaopen/foundation v0.1.0
