@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.34.2
-// 	protoc        v5.28.2
+// 	protoc-gen-go v1.36.1
+// 	protoc        v5.29.3
 // source: method_options.proto
 
 package proto
