@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.34.2
-// 	protoc        v5.28.2
+// 	protoc-gen-go v1.36.1
+// 	protoc        v5.29.3
 // source: batch.proto
 
 package proto
@@ -70,28 +70,25 @@ func (KeyType) EnumDescriptor() ([]byte, []int) {
 }
 
 type MultiSwap struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Id            []byte                 `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
+	Creator       []byte                 `protobuf:"bytes,2,opt,name=creator,proto3" json:"creator,omitempty"`
+	Owner         []byte                 `protobuf:"bytes,3,opt,name=owner,proto3" json:"owner,omitempty"`
+	Token         string                 `protobuf:"bytes,4,opt,name=token,proto3" json:"token,omitempty"`
+	From          string                 `protobuf:"bytes,5,opt,name=from,proto3" json:"from,omitempty"`
+	To            string                 `protobuf:"bytes,6,opt,name=to,proto3" json:"to,omitempty"`
+	Hash          []byte                 `protobuf:"bytes,7,opt,name=hash,proto3" json:"hash,omitempty"`
+	Timeout       int64                  `protobuf:"varint,8,opt,name=timeout,proto3" json:"timeout,omitempty"`
+	Assets        []*Asset               `protobuf:"bytes,9,rep,name=assets,proto3" json:"assets,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Id      []byte   `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
-	Creator []byte   `protobuf:"bytes,2,opt,name=creator,proto3" json:"creator,omitempty"`
-	Owner   []byte   `protobuf:"bytes,3,opt,name=owner,proto3" json:"owner,omitempty"`
-	Token   string   `protobuf:"bytes,4,opt,name=token,proto3" json:"token,omitempty"`
-	From    string   `protobuf:"bytes,5,opt,name=from,proto3" json:"from,omitempty"`
-	To      string   `protobuf:"bytes,6,opt,name=to,proto3" json:"to,omitempty"`
-	Hash    []byte   `protobuf:"bytes,7,opt,name=hash,proto3" json:"hash,omitempty"`
-	Timeout int64    `protobuf:"varint,8,opt,name=timeout,proto3" json:"timeout,omitempty"`
-	Assets  []*Asset `protobuf:"bytes,9,rep,name=assets,proto3" json:"assets,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MultiSwap) Reset() {
 	*x = MultiSwap{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MultiSwap) String() string {
@@ -102,7 +99,7 @@ func (*MultiSwap) ProtoMessage() {}
 
 func (x *MultiSwap) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -181,21 +178,18 @@ func (x *MultiSwap) GetAssets() []*Asset {
 }
 
 type Asset struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Group         string                 `protobuf:"bytes,1,opt,name=group,proto3" json:"group,omitempty"`
+	Amount        []byte                 `protobuf:"bytes,2,opt,name=amount,proto3" json:"amount,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Group  string `protobuf:"bytes,1,opt,name=group,proto3" json:"group,omitempty"`
-	Amount []byte `protobuf:"bytes,2,opt,name=amount,proto3" json:"amount,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Asset) Reset() {
 	*x = Asset{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Asset) String() string {
@@ -206,7 +200,7 @@ func (*Asset) ProtoMessage() {}
 
 func (x *Asset) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -236,28 +230,25 @@ func (x *Asset) GetAmount() []byte {
 }
 
 type Swap struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Id            []byte                 `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
+	Creator       []byte                 `protobuf:"bytes,2,opt,name=creator,proto3" json:"creator,omitempty"`
+	Owner         []byte                 `protobuf:"bytes,3,opt,name=owner,proto3" json:"owner,omitempty"`
+	Token         string                 `protobuf:"bytes,4,opt,name=token,proto3" json:"token,omitempty"`
+	Amount        []byte                 `protobuf:"bytes,5,opt,name=amount,proto3" json:"amount,omitempty"`
+	From          string                 `protobuf:"bytes,6,opt,name=from,proto3" json:"from,omitempty"`
+	To            string                 `protobuf:"bytes,7,opt,name=to,proto3" json:"to,omitempty"`
+	Hash          []byte                 `protobuf:"bytes,8,opt,name=hash,proto3" json:"hash,omitempty"`
+	Timeout       int64                  `protobuf:"varint,9,opt,name=timeout,proto3" json:"timeout,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Id      []byte `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
-	Creator []byte `protobuf:"bytes,2,opt,name=creator,proto3" json:"creator,omitempty"`
-	Owner   []byte `protobuf:"bytes,3,opt,name=owner,proto3" json:"owner,omitempty"`
-	Token   string `protobuf:"bytes,4,opt,name=token,proto3" json:"token,omitempty"`
-	Amount  []byte `protobuf:"bytes,5,opt,name=amount,proto3" json:"amount,omitempty"`
-	From    string `protobuf:"bytes,6,opt,name=from,proto3" json:"from,omitempty"`
-	To      string `protobuf:"bytes,7,opt,name=to,proto3" json:"to,omitempty"`
-	Hash    []byte `protobuf:"bytes,8,opt,name=hash,proto3" json:"hash,omitempty"`
-	Timeout int64  `protobuf:"varint,9,opt,name=timeout,proto3" json:"timeout,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Swap) Reset() {
 	*x = Swap{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Swap) String() string {
@@ -268,7 +259,7 @@ func (*Swap) ProtoMessage() {}
 
 func (x *Swap) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -347,21 +338,18 @@ func (x *Swap) GetTimeout() int64 {
 }
 
 type SwapKey struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Id            []byte                 `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
+	Key           string                 `protobuf:"bytes,2,opt,name=key,proto3" json:"key,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Id  []byte `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
-	Key string `protobuf:"bytes,2,opt,name=key,proto3" json:"key,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *SwapKey) Reset() {
 	*x = SwapKey{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SwapKey) String() string {
@@ -372,7 +360,7 @@ func (*SwapKey) ProtoMessage() {}
 
 func (x *SwapKey) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -402,24 +390,21 @@ func (x *SwapKey) GetKey() string {
 }
 
 type Batch struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	TxIDs          [][]byte     `protobuf:"bytes,1,rep,name=txIDs,proto3" json:"txIDs,omitempty"`
-	Swaps          []*Swap      `protobuf:"bytes,2,rep,name=swaps,proto3" json:"swaps,omitempty"`
-	Keys           []*SwapKey   `protobuf:"bytes,3,rep,name=keys,proto3" json:"keys,omitempty"`
-	MultiSwapsKeys []*SwapKey   `protobuf:"bytes,4,rep,name=multi_swaps_keys,json=multiSwapsKeys,proto3" json:"multi_swaps_keys,omitempty"`
-	MultiSwaps     []*MultiSwap `protobuf:"bytes,5,rep,name=multi_swaps,json=multiSwaps,proto3" json:"multi_swaps,omitempty"`
+	state          protoimpl.MessageState `protogen:"open.v1"`
+	TxIDs          [][]byte               `protobuf:"bytes,1,rep,name=txIDs,proto3" json:"txIDs,omitempty"`
+	Swaps          []*Swap                `protobuf:"bytes,2,rep,name=swaps,proto3" json:"swaps,omitempty"`
+	Keys           []*SwapKey             `protobuf:"bytes,3,rep,name=keys,proto3" json:"keys,omitempty"`
+	MultiSwapsKeys []*SwapKey             `protobuf:"bytes,4,rep,name=multi_swaps_keys,json=multiSwapsKeys,proto3" json:"multi_swaps_keys,omitempty"`
+	MultiSwaps     []*MultiSwap           `protobuf:"bytes,5,rep,name=multi_swaps,json=multiSwaps,proto3" json:"multi_swaps,omitempty"`
+	unknownFields  protoimpl.UnknownFields
+	sizeCache      protoimpl.SizeCache
 }
 
 func (x *Batch) Reset() {
 	*x = Batch{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Batch) String() string {
@@ -430,7 +415,7 @@ func (*Batch) ProtoMessage() {}
 
 func (x *Batch) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -481,22 +466,19 @@ func (x *Batch) GetMultiSwaps() []*MultiSwap {
 }
 
 type InitArgs struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	PlatformSKI   []byte                 `protobuf:"bytes,1,opt,name=platformSKI,proto3" json:"platformSKI,omitempty"`
+	RobotSKI      []byte                 `protobuf:"bytes,2,opt,name=robotSKI,proto3" json:"robotSKI,omitempty"`
+	Args          []string               `protobuf:"bytes,3,rep,name=args,proto3" json:"args,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	PlatformSKI []byte   `protobuf:"bytes,1,opt,name=platformSKI,proto3" json:"platformSKI,omitempty"`
-	RobotSKI    []byte   `protobuf:"bytes,2,opt,name=robotSKI,proto3" json:"robotSKI,omitempty"`
-	Args        []string `protobuf:"bytes,3,rep,name=args,proto3" json:"args,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *InitArgs) Reset() {
 	*x = InitArgs{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *InitArgs) String() string {
@@ -507,7 +489,7 @@ func (*InitArgs) ProtoMessage() {}
 
 func (x *InitArgs) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -544,22 +526,19 @@ func (x *InitArgs) GetArgs() []string {
 }
 
 type WriteElement struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Key           string                 `protobuf:"bytes,1,opt,name=key,proto3" json:"key,omitempty"`
+	Value         []byte                 `protobuf:"bytes,2,opt,name=value,proto3" json:"value,omitempty"`
+	IsDeleted     bool                   `protobuf:"varint,3,opt,name=is_deleted,json=isDeleted,proto3" json:"is_deleted,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Key       string `protobuf:"bytes,1,opt,name=key,proto3" json:"key,omitempty"`
-	Value     []byte `protobuf:"bytes,2,opt,name=value,proto3" json:"value,omitempty"`
-	IsDeleted bool   `protobuf:"varint,3,opt,name=is_deleted,json=isDeleted,proto3" json:"is_deleted,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *WriteElement) Reset() {
 	*x = WriteElement{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[6]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[6]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *WriteElement) String() string {
@@ -570,7 +549,7 @@ func (*WriteElement) ProtoMessage() {}
 
 func (x *WriteElement) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[6]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -607,21 +586,18 @@ func (x *WriteElement) GetIsDeleted() bool {
 }
 
 type ResponseError struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Code          int32                  `protobuf:"varint,1,opt,name=code,proto3" json:"code,omitempty"`
+	Error         string                 `protobuf:"bytes,2,opt,name=error,proto3" json:"error,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Code  int32  `protobuf:"varint,1,opt,name=code,proto3" json:"code,omitempty"`
-	Error string `protobuf:"bytes,2,opt,name=error,proto3" json:"error,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *ResponseError) Reset() {
 	*x = ResponseError{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[7]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[7]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ResponseError) String() string {
@@ -632,7 +608,7 @@ func (*ResponseError) ProtoMessage() {}
 
 func (x *ResponseError) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[7]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -662,22 +638,19 @@ func (x *ResponseError) GetError() string {
 }
 
 type SwapResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Id            []byte                 `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
+	Error         *ResponseError         `protobuf:"bytes,2,opt,name=error,proto3" json:"error,omitempty"`
+	Writes        []*WriteElement        `protobuf:"bytes,3,rep,name=writes,proto3" json:"writes,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Id     []byte          `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
-	Error  *ResponseError  `protobuf:"bytes,2,opt,name=error,proto3" json:"error,omitempty"`
-	Writes []*WriteElement `protobuf:"bytes,3,rep,name=writes,proto3" json:"writes,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *SwapResponse) Reset() {
 	*x = SwapResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[8]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[8]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SwapResponse) String() string {
@@ -688,7 +661,7 @@ func (*SwapResponse) ProtoMessage() {}
 
 func (x *SwapResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[8]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -725,26 +698,23 @@ func (x *SwapResponse) GetWrites() []*WriteElement {
 }
 
 type AccountingRecord struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	Token                string `protobuf:"bytes,1,opt,name=token,proto3" json:"token,omitempty"`
-	Sender               []byte `protobuf:"bytes,2,opt,name=sender,proto3" json:"sender,omitempty"`
-	Recipient            []byte `protobuf:"bytes,3,opt,name=recipient,proto3" json:"recipient,omitempty"`
-	Amount               []byte `protobuf:"bytes,4,opt,name=amount,proto3" json:"amount,omitempty"`
-	Reason               string `protobuf:"bytes,5,opt,name=reason,proto3" json:"reason,omitempty"`
-	SenderBalanceType    int32  `protobuf:"varint,6,opt,name=senderBalanceType,proto3" json:"senderBalanceType,omitempty"`
-	RecipientBalanceType int32  `protobuf:"varint,7,opt,name=recipientBalanceType,proto3" json:"recipientBalanceType,omitempty"`
+	state                protoimpl.MessageState `protogen:"open.v1"`
+	Token                string                 `protobuf:"bytes,1,opt,name=token,proto3" json:"token,omitempty"`
+	Sender               []byte                 `protobuf:"bytes,2,opt,name=sender,proto3" json:"sender,omitempty"`
+	Recipient            []byte                 `protobuf:"bytes,3,opt,name=recipient,proto3" json:"recipient,omitempty"`
+	Amount               []byte                 `protobuf:"bytes,4,opt,name=amount,proto3" json:"amount,omitempty"`
+	Reason               string                 `protobuf:"bytes,5,opt,name=reason,proto3" json:"reason,omitempty"`
+	SenderBalanceType    int32                  `protobuf:"varint,6,opt,name=senderBalanceType,proto3" json:"senderBalanceType,omitempty"`
+	RecipientBalanceType int32                  `protobuf:"varint,7,opt,name=recipientBalanceType,proto3" json:"recipientBalanceType,omitempty"`
+	unknownFields        protoimpl.UnknownFields
+	sizeCache            protoimpl.SizeCache
 }
 
 func (x *AccountingRecord) Reset() {
 	*x = AccountingRecord{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[9]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[9]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *AccountingRecord) String() string {
@@ -755,7 +725,7 @@ func (*AccountingRecord) ProtoMessage() {}
 
 func (x *AccountingRecord) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[9]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -820,21 +790,18 @@ func (x *AccountingRecord) GetRecipientBalanceType() int32 {
 }
 
 type Event struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Name          string                 `protobuf:"bytes,1,opt,name=name,proto3" json:"name,omitempty"`
+	Value         []byte                 `protobuf:"bytes,2,opt,name=value,proto3" json:"value,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Name  string `protobuf:"bytes,1,opt,name=name,proto3" json:"name,omitempty"`
-	Value []byte `protobuf:"bytes,2,opt,name=value,proto3" json:"value,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Event) Reset() {
 	*x = Event{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[10]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[10]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Event) String() string {
@@ -845,7 +812,7 @@ func (*Event) ProtoMessage() {}
 
 func (x *Event) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[10]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -875,23 +842,20 @@ func (x *Event) GetValue() []byte {
 }
 
 type TxResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Id            []byte                 `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
+	Method        string                 `protobuf:"bytes,2,opt,name=method,proto3" json:"method,omitempty"`
+	Error         *ResponseError         `protobuf:"bytes,3,opt,name=error,proto3" json:"error,omitempty"`
+	Writes        []*WriteElement        `protobuf:"bytes,4,rep,name=writes,proto3" json:"writes,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Id     []byte          `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
-	Method string          `protobuf:"bytes,2,opt,name=method,proto3" json:"method,omitempty"`
-	Error  *ResponseError  `protobuf:"bytes,3,opt,name=error,proto3" json:"error,omitempty"`
-	Writes []*WriteElement `protobuf:"bytes,4,rep,name=writes,proto3" json:"writes,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *TxResponse) Reset() {
 	*x = TxResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[11]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[11]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TxResponse) String() string {
@@ -902,7 +866,7 @@ func (*TxResponse) ProtoMessage() {}
 
 func (x *TxResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[11]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -946,25 +910,22 @@ func (x *TxResponse) GetWrites() []*WriteElement {
 }
 
 type BatchTxEvent struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Id            []byte                 `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
+	Method        string                 `protobuf:"bytes,2,opt,name=method,proto3" json:"method,omitempty"`
+	Error         *ResponseError         `protobuf:"bytes,3,opt,name=error,proto3" json:"error,omitempty"`
+	Result        []byte                 `protobuf:"bytes,4,opt,name=result,proto3" json:"result,omitempty"`
+	Events        []*Event               `protobuf:"bytes,6,rep,name=events,proto3" json:"events,omitempty"`
+	Accounting    []*AccountingRecord    `protobuf:"bytes,7,rep,name=accounting,proto3" json:"accounting,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Id         []byte              `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
-	Method     string              `protobuf:"bytes,2,opt,name=method,proto3" json:"method,omitempty"`
-	Error      *ResponseError      `protobuf:"bytes,3,opt,name=error,proto3" json:"error,omitempty"`
-	Result     []byte              `protobuf:"bytes,4,opt,name=result,proto3" json:"result,omitempty"`
-	Events     []*Event            `protobuf:"bytes,6,rep,name=events,proto3" json:"events,omitempty"`
-	Accounting []*AccountingRecord `protobuf:"bytes,7,rep,name=accounting,proto3" json:"accounting,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *BatchTxEvent) Reset() {
 	*x = BatchTxEvent{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[12]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[12]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *BatchTxEvent) String() string {
@@ -975,7 +936,7 @@ func (*BatchTxEvent) ProtoMessage() {}
 
 func (x *BatchTxEvent) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[12]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1033,20 +994,17 @@ func (x *BatchTxEvent) GetAccounting() []*AccountingRecord {
 }
 
 type BatchEvent struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Events        []*BatchTxEvent        `protobuf:"bytes,1,rep,name=events,proto3" json:"events,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Events []*BatchTxEvent `protobuf:"bytes,1,rep,name=events,proto3" json:"events,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *BatchEvent) Reset() {
 	*x = BatchEvent{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[13]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[13]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *BatchEvent) String() string {
@@ -1057,7 +1015,7 @@ func (*BatchEvent) ProtoMessage() {}
 
 func (x *BatchEvent) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[13]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1080,24 +1038,21 @@ func (x *BatchEvent) GetEvents() []*BatchTxEvent {
 }
 
 type BatchResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	TxResponses      []*TxResponse   `protobuf:"bytes,1,rep,name=tx_responses,json=txResponses,proto3" json:"tx_responses,omitempty"`
-	CreatedSwaps     []*Swap         `protobuf:"bytes,2,rep,name=created_swaps,json=createdSwaps,proto3" json:"created_swaps,omitempty"`
-	SwapResponses    []*SwapResponse `protobuf:"bytes,3,rep,name=swap_responses,json=swapResponses,proto3" json:"swap_responses,omitempty"`
-	SwapKeyResponses []*SwapResponse `protobuf:"bytes,4,rep,name=swap_key_responses,json=swapKeyResponses,proto3" json:"swap_key_responses,omitempty"`
-	CreatedMultiSwap []*MultiSwap    `protobuf:"bytes,5,rep,name=created_multi_swap,json=createdMultiSwap,proto3" json:"created_multi_swap,omitempty"`
+	state            protoimpl.MessageState `protogen:"open.v1"`
+	TxResponses      []*TxResponse          `protobuf:"bytes,1,rep,name=tx_responses,json=txResponses,proto3" json:"tx_responses,omitempty"`
+	CreatedSwaps     []*Swap                `protobuf:"bytes,2,rep,name=created_swaps,json=createdSwaps,proto3" json:"created_swaps,omitempty"`
+	SwapResponses    []*SwapResponse        `protobuf:"bytes,3,rep,name=swap_responses,json=swapResponses,proto3" json:"swap_responses,omitempty"`
+	SwapKeyResponses []*SwapResponse        `protobuf:"bytes,4,rep,name=swap_key_responses,json=swapKeyResponses,proto3" json:"swap_key_responses,omitempty"`
+	CreatedMultiSwap []*MultiSwap           `protobuf:"bytes,5,rep,name=created_multi_swap,json=createdMultiSwap,proto3" json:"created_multi_swap,omitempty"`
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
 }
 
 func (x *BatchResponse) Reset() {
 	*x = BatchResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[14]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[14]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *BatchResponse) String() string {
@@ -1108,7 +1063,7 @@ func (*BatchResponse) ProtoMessage() {}
 
 func (x *BatchResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[14]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1159,20 +1114,17 @@ func (x *BatchResponse) GetCreatedMultiSwap() []*MultiSwap {
 }
 
 type Nested struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Args          []string               `protobuf:"bytes,1,rep,name=args,proto3" json:"args,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Args []string `protobuf:"bytes,1,rep,name=args,proto3" json:"args,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Nested) Reset() {
 	*x = Nested{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[15]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[15]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Nested) String() string {
@@ -1183,7 +1135,7 @@ func (*Nested) ProtoMessage() {}
 
 func (x *Nested) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[15]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1206,23 +1158,20 @@ func (x *Nested) GetArgs() []string {
 }
 
 type TokenFee struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Currency      string                 `protobuf:"bytes,1,opt,name=currency,proto3" json:"currency,omitempty"`
+	Fee           []byte                 `protobuf:"bytes,2,opt,name=fee,proto3" json:"fee,omitempty"`
+	Floor         []byte                 `protobuf:"bytes,3,opt,name=floor,proto3" json:"floor,omitempty"`
+	Cap           []byte                 `protobuf:"bytes,4,opt,name=cap,proto3" json:"cap,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Currency string `protobuf:"bytes,1,opt,name=currency,proto3" json:"currency,omitempty"`
-	Fee      []byte `protobuf:"bytes,2,opt,name=fee,proto3" json:"fee,omitempty"`
-	Floor    []byte `protobuf:"bytes,3,opt,name=floor,proto3" json:"floor,omitempty"`
-	Cap      []byte `protobuf:"bytes,4,opt,name=cap,proto3" json:"cap,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *TokenFee) Reset() {
 	*x = TokenFee{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[16]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[16]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TokenFee) String() string {
@@ -1233,7 +1182,7 @@ func (*TokenFee) ProtoMessage() {}
 
 func (x *TokenFee) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[16]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1277,25 +1226,22 @@ func (x *TokenFee) GetCap() []byte {
 }
 
 type TokenRate struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	DealType      string                 `protobuf:"bytes,1,opt,name=deal_type,json=dealType,proto3" json:"deal_type,omitempty"`
+	Currency      string                 `protobuf:"bytes,2,opt,name=currency,proto3" json:"currency,omitempty"`
+	Rate          []byte                 `protobuf:"bytes,3,opt,name=rate,proto3" json:"rate,omitempty"`
+	Min           []byte                 `protobuf:"bytes,4,opt,name=min,proto3" json:"min,omitempty"`
+	Max           []byte                 `protobuf:"bytes,5,opt,name=max,proto3" json:"max,omitempty"`
+	Issuer        *Address               `protobuf:"bytes,6,opt,name=issuer,proto3" json:"issuer,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	DealType string   `protobuf:"bytes,1,opt,name=deal_type,json=dealType,proto3" json:"deal_type,omitempty"`
-	Currency string   `protobuf:"bytes,2,opt,name=currency,proto3" json:"currency,omitempty"`
-	Rate     []byte   `protobuf:"bytes,3,opt,name=rate,proto3" json:"rate,omitempty"`
-	Min      []byte   `protobuf:"bytes,4,opt,name=min,proto3" json:"min,omitempty"`
-	Max      []byte   `protobuf:"bytes,5,opt,name=max,proto3" json:"max,omitempty"`
-	Issuer   *Address `protobuf:"bytes,6,opt,name=issuer,proto3" json:"issuer,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *TokenRate) Reset() {
 	*x = TokenRate{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[17]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[17]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TokenRate) String() string {
@@ -1306,7 +1252,7 @@ func (*TokenRate) ProtoMessage() {}
 
 func (x *TokenRate) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[17]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1364,23 +1310,20 @@ func (x *TokenRate) GetIssuer() *Address {
 }
 
 type Token struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	TotalEmission []byte                 `protobuf:"bytes,1,opt,name=total_emission,json=totalEmission,proto3" json:"total_emission,omitempty"`
+	Fee           *TokenFee              `protobuf:"bytes,2,opt,name=fee,proto3" json:"fee,omitempty"`
+	Rates         []*TokenRate           `protobuf:"bytes,3,rep,name=rates,proto3" json:"rates,omitempty"`
+	FeeAddress    []byte                 `protobuf:"bytes,4,opt,name=fee_address,json=feeAddress,proto3" json:"fee_address,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	TotalEmission []byte       `protobuf:"bytes,1,opt,name=total_emission,json=totalEmission,proto3" json:"total_emission,omitempty"`
-	Fee           *TokenFee    `protobuf:"bytes,2,opt,name=fee,proto3" json:"fee,omitempty"`
-	Rates         []*TokenRate `protobuf:"bytes,3,rep,name=rates,proto3" json:"rates,omitempty"`
-	FeeAddress    []byte       `protobuf:"bytes,4,opt,name=fee_address,json=feeAddress,proto3" json:"fee_address,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Token) Reset() {
 	*x = Token{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[18]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[18]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Token) String() string {
@@ -1391,7 +1334,7 @@ func (*Token) ProtoMessage() {}
 
 func (x *Token) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[18]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1435,20 +1378,17 @@ func (x *Token) GetFeeAddress() []byte {
 }
 
 type HaveRight struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	HaveRight     bool                   `protobuf:"varint,1,opt,name=haveRight,proto3" json:"haveRight,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	HaveRight bool `protobuf:"varint,1,opt,name=haveRight,proto3" json:"haveRight,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *HaveRight) Reset() {
 	*x = HaveRight{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[19]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[19]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *HaveRight) String() string {
@@ -1459,7 +1399,7 @@ func (*HaveRight) ProtoMessage() {}
 
 func (x *HaveRight) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[19]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1482,25 +1422,22 @@ func (x *HaveRight) GetHaveRight() bool {
 }
 
 type Right struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	ChannelName   string                 `protobuf:"bytes,1,opt,name=channelName,proto3" json:"channelName,omitempty"`
+	ChaincodeName string                 `protobuf:"bytes,2,opt,name=chaincodeName,proto3" json:"chaincodeName,omitempty"`
+	RoleName      string                 `protobuf:"bytes,3,opt,name=roleName,proto3" json:"roleName,omitempty"`
+	OperationName string                 `protobuf:"bytes,4,opt,name=operationName,proto3" json:"operationName,omitempty"`
+	Address       *Address               `protobuf:"bytes,5,opt,name=address,proto3" json:"address,omitempty"`
+	HaveRight     *HaveRight             `protobuf:"bytes,6,opt,name=haveRight,proto3" json:"haveRight,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	ChannelName   string     `protobuf:"bytes,1,opt,name=channelName,proto3" json:"channelName,omitempty"`
-	ChaincodeName string     `protobuf:"bytes,2,opt,name=chaincodeName,proto3" json:"chaincodeName,omitempty"`
-	RoleName      string     `protobuf:"bytes,3,opt,name=roleName,proto3" json:"roleName,omitempty"`
-	OperationName string     `protobuf:"bytes,4,opt,name=operationName,proto3" json:"operationName,omitempty"`
-	Address       *Address   `protobuf:"bytes,5,opt,name=address,proto3" json:"address,omitempty"`
-	HaveRight     *HaveRight `protobuf:"bytes,6,opt,name=haveRight,proto3" json:"haveRight,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Right) Reset() {
 	*x = Right{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[20]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[20]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Right) String() string {
@@ -1511,7 +1448,7 @@ func (*Right) ProtoMessage() {}
 
 func (x *Right) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[20]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1569,21 +1506,18 @@ func (x *Right) GetHaveRight() *HaveRight {
 }
 
 type AccountRights struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Address       *Address               `protobuf:"bytes,1,opt,name=address,proto3" json:"address,omitempty"`
+	Rights        []*Right               `protobuf:"bytes,2,rep,name=rights,proto3" json:"rights,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Address *Address `protobuf:"bytes,1,opt,name=address,proto3" json:"address,omitempty"`
-	Rights  []*Right `protobuf:"bytes,2,rep,name=rights,proto3" json:"rights,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *AccountRights) Reset() {
 	*x = AccountRights{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[21]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[21]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *AccountRights) String() string {
@@ -1594,7 +1528,7 @@ func (*AccountRights) ProtoMessage() {}
 
 func (x *AccountRights) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[21]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1624,20 +1558,17 @@ func (x *AccountRights) GetRights() []*Right {
 }
 
 type Accounts struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Addresses     []*Address             `protobuf:"bytes,1,rep,name=addresses,proto3" json:"addresses,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Addresses []*Address `protobuf:"bytes,1,rep,name=addresses,proto3" json:"addresses,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Accounts) Reset() {
 	*x = Accounts{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[22]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[22]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Accounts) String() string {
@@ -1648,7 +1579,7 @@ func (*Accounts) ProtoMessage() {}
 
 func (x *Accounts) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[22]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1671,20 +1602,17 @@ func (x *Accounts) GetAddresses() []*Address {
 }
 
 type Operations struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Operations    []string               `protobuf:"bytes,1,rep,name=operations,proto3" json:"operations,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Operations []string `protobuf:"bytes,1,rep,name=operations,proto3" json:"operations,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Operations) Reset() {
 	*x = Operations{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[23]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[23]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Operations) String() string {
@@ -1695,7 +1623,7 @@ func (*Operations) ProtoMessage() {}
 
 func (x *Operations) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[23]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1718,21 +1646,18 @@ func (x *Operations) GetOperations() []string {
 }
 
 type OperationRights struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	OperationName string                 `protobuf:"bytes,1,opt,name=operationName,proto3" json:"operationName,omitempty"`
+	Rights        []*Right               `protobuf:"bytes,2,rep,name=rights,proto3" json:"rights,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	OperationName string   `protobuf:"bytes,1,opt,name=operationName,proto3" json:"operationName,omitempty"`
-	Rights        []*Right `protobuf:"bytes,2,rep,name=rights,proto3" json:"rights,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *OperationRights) Reset() {
 	*x = OperationRights{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[24]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[24]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *OperationRights) String() string {
@@ -1743,7 +1668,7 @@ func (*OperationRights) ProtoMessage() {}
 
 func (x *OperationRights) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[24]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1773,24 +1698,21 @@ func (x *OperationRights) GetRights() []*Right {
 }
 
 type Industrial struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Groups        []*IndustrialGroup     `protobuf:"bytes,1,rep,name=groups,proto3" json:"groups,omitempty"`
+	Fee           *TokenFee              `protobuf:"bytes,2,opt,name=fee,proto3" json:"fee,omitempty"`
+	Rates         []*TokenRate           `protobuf:"bytes,3,rep,name=rates,proto3" json:"rates,omitempty"`
+	FeeAddress    []byte                 `protobuf:"bytes,4,opt,name=fee_address,json=feeAddress,proto3" json:"fee_address,omitempty"`
+	Initialized   bool                   `protobuf:"varint,5,opt,name=initialized,proto3" json:"initialized,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Groups      []*IndustrialGroup `protobuf:"bytes,1,rep,name=groups,proto3" json:"groups,omitempty"`
-	Fee         *TokenFee          `protobuf:"bytes,2,opt,name=fee,proto3" json:"fee,omitempty"`
-	Rates       []*TokenRate       `protobuf:"bytes,3,rep,name=rates,proto3" json:"rates,omitempty"`
-	FeeAddress  []byte             `protobuf:"bytes,4,opt,name=fee_address,json=feeAddress,proto3" json:"fee_address,omitempty"`
-	Initialized bool               `protobuf:"varint,5,opt,name=initialized,proto3" json:"initialized,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Industrial) Reset() {
 	*x = Industrial{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[25]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[25]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Industrial) String() string {
@@ -1801,7 +1723,7 @@ func (*Industrial) ProtoMessage() {}
 
 func (x *Industrial) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[25]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1852,23 +1774,20 @@ func (x *Industrial) GetInitialized() bool {
 }
 
 type IndustrialGroup struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Id            string                 `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
+	Emission      []byte                 `protobuf:"bytes,2,opt,name=emission,proto3" json:"emission,omitempty"`
+	Maturity      int64                  `protobuf:"varint,3,opt,name=maturity,proto3" json:"maturity,omitempty"`
+	Note          string                 `protobuf:"bytes,4,opt,name=note,proto3" json:"note,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Id       string `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
-	Emission []byte `protobuf:"bytes,2,opt,name=emission,proto3" json:"emission,omitempty"`
-	Maturity int64  `protobuf:"varint,3,opt,name=maturity,proto3" json:"maturity,omitempty"`
-	Note     string `protobuf:"bytes,4,opt,name=note,proto3" json:"note,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *IndustrialGroup) Reset() {
 	*x = IndustrialGroup{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[26]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[26]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *IndustrialGroup) String() string {
@@ -1879,7 +1798,7 @@ func (*IndustrialGroup) ProtoMessage() {}
 
 func (x *IndustrialGroup) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[26]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1923,22 +1842,19 @@ func (x *IndustrialGroup) GetNote() string {
 }
 
 type AccountInfo struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	KycHash       string                 `protobuf:"bytes,1,opt,name=kycHash,proto3" json:"kycHash,omitempty"`
+	GrayListed    bool                   `protobuf:"varint,2,opt,name=grayListed,proto3" json:"grayListed,omitempty"`
+	BlackListed   bool                   `protobuf:"varint,3,opt,name=blackListed,proto3" json:"blackListed,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	KycHash     string `protobuf:"bytes,1,opt,name=kycHash,proto3" json:"kycHash,omitempty"`
-	GrayListed  bool   `protobuf:"varint,2,opt,name=grayListed,proto3" json:"grayListed,omitempty"`
-	BlackListed bool   `protobuf:"varint,3,opt,name=blackListed,proto3" json:"blackListed,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *AccountInfo) Reset() {
 	*x = AccountInfo{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[27]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[27]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *AccountInfo) String() string {
@@ -1949,7 +1865,7 @@ func (*AccountInfo) ProtoMessage() {}
 
 func (x *AccountInfo) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[27]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1986,23 +1902,20 @@ func (x *AccountInfo) GetBlackListed() bool {
 }
 
 type Address struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	UserID        string                 `protobuf:"bytes,1,opt,name=userID,proto3" json:"userID,omitempty"`
+	Address       []byte                 `protobuf:"bytes,2,opt,name=address,proto3" json:"address,omitempty"`
+	IsIndustrial  bool                   `protobuf:"varint,3,opt,name=isIndustrial,proto3" json:"isIndustrial,omitempty"`
+	IsMultisig    bool                   `protobuf:"varint,4,opt,name=isMultisig,proto3" json:"isMultisig,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	UserID       string `protobuf:"bytes,1,opt,name=userID,proto3" json:"userID,omitempty"`
-	Address      []byte `protobuf:"bytes,2,opt,name=address,proto3" json:"address,omitempty"`
-	IsIndustrial bool   `protobuf:"varint,3,opt,name=isIndustrial,proto3" json:"isIndustrial,omitempty"`
-	IsMultisig   bool   `protobuf:"varint,4,opt,name=isMultisig,proto3" json:"isMultisig,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Address) Reset() {
 	*x = Address{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[28]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[28]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Address) String() string {
@@ -2013,7 +1926,7 @@ func (*Address) ProtoMessage() {}
 
 func (x *Address) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[28]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2057,21 +1970,18 @@ func (x *Address) GetIsMultisig() bool {
 }
 
 type AdditionalKey struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	PublicKeyBase58 string   `protobuf:"bytes,1,opt,name=publicKeyBase58,proto3" json:"publicKeyBase58,omitempty"`
-	Labels          []string `protobuf:"bytes,2,rep,name=labels,proto3" json:"labels,omitempty"`
+	state           protoimpl.MessageState `protogen:"open.v1"`
+	PublicKeyBase58 string                 `protobuf:"bytes,1,opt,name=publicKeyBase58,proto3" json:"publicKeyBase58,omitempty"`
+	Labels          []string               `protobuf:"bytes,2,rep,name=labels,proto3" json:"labels,omitempty"`
+	unknownFields   protoimpl.UnknownFields
+	sizeCache       protoimpl.SizeCache
 }
 
 func (x *AdditionalKey) Reset() {
 	*x = AdditionalKey{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[29]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[29]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *AdditionalKey) String() string {
@@ -2082,7 +1992,7 @@ func (*AdditionalKey) ProtoMessage() {}
 
 func (x *AdditionalKey) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[29]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2112,25 +2022,22 @@ func (x *AdditionalKey) GetLabels() []string {
 }
 
 type SignedAddress struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	Address         *Address         `protobuf:"bytes,1,opt,name=address,proto3" json:"address,omitempty"`
-	SignedTx        []string         `protobuf:"bytes,2,rep,name=signedTx,proto3" json:"signedTx,omitempty"`
-	SignaturePolicy *SignaturePolicy `protobuf:"bytes,5,opt,name=signaturePolicy,proto3" json:"signaturePolicy,omitempty"`
-	Reason          string           `protobuf:"bytes,6,opt,name=reason,proto3" json:"reason,omitempty"`
-	ReasonId        int32            `protobuf:"varint,7,opt,name=reasonId,proto3" json:"reasonId,omitempty"`
-	AdditionalKeys  []*AdditionalKey `protobuf:"bytes,8,rep,name=additionalKeys,proto3" json:"additionalKeys,omitempty"`
+	state           protoimpl.MessageState `protogen:"open.v1"`
+	Address         *Address               `protobuf:"bytes,1,opt,name=address,proto3" json:"address,omitempty"`
+	SignedTx        []string               `protobuf:"bytes,2,rep,name=signedTx,proto3" json:"signedTx,omitempty"`
+	SignaturePolicy *SignaturePolicy       `protobuf:"bytes,5,opt,name=signaturePolicy,proto3" json:"signaturePolicy,omitempty"`
+	Reason          string                 `protobuf:"bytes,6,opt,name=reason,proto3" json:"reason,omitempty"`
+	ReasonId        int32                  `protobuf:"varint,7,opt,name=reasonId,proto3" json:"reasonId,omitempty"`
+	AdditionalKeys  []*AdditionalKey       `protobuf:"bytes,8,rep,name=additionalKeys,proto3" json:"additionalKeys,omitempty"`
+	unknownFields   protoimpl.UnknownFields
+	sizeCache       protoimpl.SizeCache
 }
 
 func (x *SignedAddress) Reset() {
 	*x = SignedAddress{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[30]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[30]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SignedAddress) String() string {
@@ -2141,7 +2048,7 @@ func (*SignedAddress) ProtoMessage() {}
 
 func (x *SignedAddress) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[30]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2199,22 +2106,19 @@ func (x *SignedAddress) GetAdditionalKeys() []*AdditionalKey {
 }
 
 type SignaturePolicy struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	N                   uint32   `protobuf:"varint,1,opt,name=n,proto3" json:"n,omitempty"`
-	PubKeys             [][]byte `protobuf:"bytes,3,rep,name=pubKeys,proto3" json:"pubKeys,omitempty"`
-	ReplaceKeysSignedTx []string `protobuf:"bytes,4,rep,name=replaceKeysSignedTx,proto3" json:"replaceKeysSignedTx,omitempty"`
+	state               protoimpl.MessageState `protogen:"open.v1"`
+	N                   uint32                 `protobuf:"varint,1,opt,name=n,proto3" json:"n,omitempty"`
+	PubKeys             [][]byte               `protobuf:"bytes,3,rep,name=pubKeys,proto3" json:"pubKeys,omitempty"`
+	ReplaceKeysSignedTx []string               `protobuf:"bytes,4,rep,name=replaceKeysSignedTx,proto3" json:"replaceKeysSignedTx,omitempty"`
+	unknownFields       protoimpl.UnknownFields
+	sizeCache           protoimpl.SizeCache
 }
 
 func (x *SignaturePolicy) Reset() {
 	*x = SignaturePolicy{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[31]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[31]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SignaturePolicy) String() string {
@@ -2225,7 +2129,7 @@ func (*SignaturePolicy) ProtoMessage() {}
 
 func (x *SignaturePolicy) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[31]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2262,22 +2166,19 @@ func (x *SignaturePolicy) GetReplaceKeysSignedTx() []string {
 }
 
 type AclResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Account       *AccountInfo           `protobuf:"bytes,1,opt,name=account,proto3" json:"account,omitempty"`
+	Address       *SignedAddress         `protobuf:"bytes,2,opt,name=address,proto3" json:"address,omitempty"`
+	KeyTypes      []KeyType              `protobuf:"varint,3,rep,packed,name=keyTypes,proto3,enum=proto.KeyType" json:"keyTypes,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Account  *AccountInfo   `protobuf:"bytes,1,opt,name=account,proto3" json:"account,omitempty"`
-	Address  *SignedAddress `protobuf:"bytes,2,opt,name=address,proto3" json:"address,omitempty"`
-	KeyTypes []KeyType      `protobuf:"varint,3,rep,packed,name=keyTypes,proto3,enum=proto.KeyType" json:"keyTypes,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *AclResponse) Reset() {
 	*x = AclResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[32]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[32]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *AclResponse) String() string {
@@ -2288,7 +2189,7 @@ func (*AclResponse) ProtoMessage() {}
 
 func (x *AclResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[32]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2325,20 +2226,17 @@ func (x *AclResponse) GetKeyTypes() []KeyType {
 }
 
 type Nonce struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Nonce         []uint64               `protobuf:"varint,1,rep,packed,name=nonce,proto3" json:"nonce,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Nonce []uint64 `protobuf:"varint,1,rep,packed,name=nonce,proto3" json:"nonce,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Nonce) Reset() {
 	*x = Nonce{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[33]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[33]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Nonce) String() string {
@@ -2349,7 +2247,7 @@ func (*Nonce) ProtoMessage() {}
 
 func (x *Nonce) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[33]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2372,26 +2270,23 @@ func (x *Nonce) GetNonce() []uint64 {
 }
 
 type PendingTx struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	Method string   `protobuf:"bytes,1,opt,name=method,proto3" json:"method,omitempty"`
-	Sender *Address `protobuf:"bytes,2,opt,name=sender,proto3" json:"sender,omitempty"`
-	Args   []string `protobuf:"bytes,3,rep,name=args,proto3" json:"args,omitempty"`
+	state  protoimpl.MessageState `protogen:"open.v1"`
+	Method string                 `protobuf:"bytes,1,opt,name=method,proto3" json:"method,omitempty"`
+	Sender *Address               `protobuf:"bytes,2,opt,name=sender,proto3" json:"sender,omitempty"`
+	Args   []string               `protobuf:"bytes,3,rep,name=args,proto3" json:"args,omitempty"`
 	// bytes ______________ = 4; the field has been deleted, avoid reusing it
-	Timestamp int64   `protobuf:"varint,5,opt,name=timestamp,proto3" json:"timestamp,omitempty"`
-	Nonce     uint64  `protobuf:"varint,6,opt,name=nonce,proto3" json:"nonce,omitempty"`
-	Pairs     []*Pair `protobuf:"bytes,7,rep,name=pairs,proto3" json:"pairs,omitempty"` // key-value pairs for telemetry settings storage
+	Timestamp     int64   `protobuf:"varint,5,opt,name=timestamp,proto3" json:"timestamp,omitempty"`
+	Nonce         uint64  `protobuf:"varint,6,opt,name=nonce,proto3" json:"nonce,omitempty"`
+	Pairs         []*Pair `protobuf:"bytes,7,rep,name=pairs,proto3" json:"pairs,omitempty"` // key-value pairs for telemetry settings storage
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PendingTx) Reset() {
 	*x = PendingTx{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[34]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[34]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PendingTx) String() string {
@@ -2402,7 +2297,7 @@ func (*PendingTx) ProtoMessage() {}
 
 func (x *PendingTx) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[34]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2461,21 +2356,18 @@ func (x *PendingTx) GetPairs() []*Pair {
 
 // key-value pairs for telemetry settings storage
 type Pair struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Key           string                 `protobuf:"bytes,1,opt,name=key,proto3" json:"key,omitempty"`
+	Value         string                 `protobuf:"bytes,2,opt,name=value,proto3" json:"value,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Key   string `protobuf:"bytes,1,opt,name=key,proto3" json:"key,omitempty"`
-	Value string `protobuf:"bytes,2,opt,name=value,proto3" json:"value,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Pair) Reset() {
 	*x = Pair{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[35]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[35]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Pair) String() string {
@@ -2486,7 +2378,7 @@ func (*Pair) ProtoMessage() {}
 
 func (x *Pair) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[35]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2516,16 +2408,13 @@ func (x *Pair) GetValue() string {
 }
 
 type CCTransfer struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	Id     string `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`         // unique transfer id
-	From   string `protobuf:"bytes,2,opt,name=from,proto3" json:"from,omitempty"`     // channel from
-	To     string `protobuf:"bytes,3,opt,name=to,proto3" json:"to,omitempty"`         // channel to
-	Token  string `protobuf:"bytes,4,opt,name=token,proto3" json:"token,omitempty"`   // transfer token
-	User   []byte `protobuf:"bytes,5,opt,name=user,proto3" json:"user,omitempty"`     // token holder
-	Amount []byte `protobuf:"bytes,6,opt,name=amount,proto3" json:"amount,omitempty"` // number of tokens
+	state  protoimpl.MessageState `protogen:"open.v1"`
+	Id     string                 `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`         // unique transfer id
+	From   string                 `protobuf:"bytes,2,opt,name=from,proto3" json:"from,omitempty"`     // channel from
+	To     string                 `protobuf:"bytes,3,opt,name=to,proto3" json:"to,omitempty"`         // channel to
+	Token  string                 `protobuf:"bytes,4,opt,name=token,proto3" json:"token,omitempty"`   // transfer token
+	User   []byte                 `protobuf:"bytes,5,opt,name=user,proto3" json:"user,omitempty"`     // token holder
+	Amount []byte                 `protobuf:"bytes,6,opt,name=amount,proto3" json:"amount,omitempty"` // number of tokens
 	// Transfer direction is an additional variable made for convenience
 	// so that you don't have to calculate it every time. It is calculated 1 time when filling the structure
 	// when executing a transaction.
@@ -2539,15 +2428,15 @@ type CCTransfer struct {
 	IsCommit         bool              `protobuf:"varint,8,opt,name=isCommit,proto3" json:"isCommit,omitempty"`                            // phase 2 sign
 	TimeAsNanos      int64             `protobuf:"varint,9,opt,name=time_as_nanos,json=timeAsNanos,proto3" json:"time_as_nanos,omitempty"` // transfer creation time in nanoseconds
 	Items            []*CCTransferItem `protobuf:"bytes,10,rep,name=items,proto3" json:"items,omitempty"`                                  // transfer multiple tokens for single transfer
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
 }
 
 func (x *CCTransfer) Reset() {
 	*x = CCTransfer{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[36]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[36]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CCTransfer) String() string {
@@ -2558,7 +2447,7 @@ func (*CCTransfer) ProtoMessage() {}
 
 func (x *CCTransfer) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[36]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2644,21 +2533,18 @@ func (x *CCTransfer) GetItems() []*CCTransferItem {
 }
 
 type CCTransfers struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Bookmark      string                 `protobuf:"bytes,1,opt,name=bookmark,proto3" json:"bookmark,omitempty"`
+	Ccts          []*CCTransfer          `protobuf:"bytes,2,rep,name=ccts,proto3" json:"ccts,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Bookmark string        `protobuf:"bytes,1,opt,name=bookmark,proto3" json:"bookmark,omitempty"`
-	Ccts     []*CCTransfer `protobuf:"bytes,2,rep,name=ccts,proto3" json:"ccts,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *CCTransfers) Reset() {
 	*x = CCTransfers{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[37]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[37]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CCTransfers) String() string {
@@ -2669,7 +2555,7 @@ func (*CCTransfers) ProtoMessage() {}
 
 func (x *CCTransfers) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[37]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -2699,21 +2585,18 @@ func (x *CCTransfers) GetCcts() []*CCTransfer {
 }
 
 type CCTransferItem struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Token         string                 `protobuf:"bytes,1,opt,name=token,proto3" json:"token,omitempty"`
+	Amount        []byte                 `protobuf:"bytes,2,opt,name=amount,proto3" json:"amount,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Token  string `protobuf:"bytes,1,opt,name=token,proto3" json:"token,omitempty"`
-	Amount []byte `protobuf:"bytes,2,opt,name=amount,proto3" json:"amount,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *CCTransferItem) Reset() {
 	*x = CCTransferItem{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_batch_proto_msgTypes[38]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_batch_proto_msgTypes[38]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CCTransferItem) String() string {
@@ -2724,7 +2607,7 @@ func (*CCTransferItem) ProtoMessage() {}
 
 func (x *CCTransferItem) ProtoReflect() protoreflect.Message {
 	mi := &file_batch_proto_msgTypes[38]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -3187,476 +3070,6 @@ func init() { file_batch_proto_init() }
 func file_batch_proto_init() {
 	if File_batch_proto != nil {
 		return
-	}
-	if !protoimpl.UnsafeEnabled {
-		file_batch_proto_msgTypes[0].Exporter = func(v any, i int) any {
-			switch v := v.(*MultiSwap); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[1].Exporter = func(v any, i int) any {
-			switch v := v.(*Asset); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[2].Exporter = func(v any, i int) any {
-			switch v := v.(*Swap); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[3].Exporter = func(v any, i int) any {
-			switch v := v.(*SwapKey); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[4].Exporter = func(v any, i int) any {
-			switch v := v.(*Batch); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[5].Exporter = func(v any, i int) any {
-			switch v := v.(*InitArgs); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[6].Exporter = func(v any, i int) any {
-			switch v := v.(*WriteElement); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[7].Exporter = func(v any, i int) any {
-			switch v := v.(*ResponseError); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[8].Exporter = func(v any, i int) any {
-			switch v := v.(*SwapResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[9].Exporter = func(v any, i int) any {
-			switch v := v.(*AccountingRecord); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[10].Exporter = func(v any, i int) any {
-			switch v := v.(*Event); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[11].Exporter = func(v any, i int) any {
-			switch v := v.(*TxResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[12].Exporter = func(v any, i int) any {
-			switch v := v.(*BatchTxEvent); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[13].Exporter = func(v any, i int) any {
-			switch v := v.(*BatchEvent); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[14].Exporter = func(v any, i int) any {
-			switch v := v.(*BatchResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[15].Exporter = func(v any, i int) any {
-			switch v := v.(*Nested); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[16].Exporter = func(v any, i int) any {
-			switch v := v.(*TokenFee); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[17].Exporter = func(v any, i int) any {
-			switch v := v.(*TokenRate); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[18].Exporter = func(v any, i int) any {
-			switch v := v.(*Token); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[19].Exporter = func(v any, i int) any {
-			switch v := v.(*HaveRight); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[20].Exporter = func(v any, i int) any {
-			switch v := v.(*Right); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[21].Exporter = func(v any, i int) any {
-			switch v := v.(*AccountRights); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[22].Exporter = func(v any, i int) any {
-			switch v := v.(*Accounts); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[23].Exporter = func(v any, i int) any {
-			switch v := v.(*Operations); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[24].Exporter = func(v any, i int) any {
-			switch v := v.(*OperationRights); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[25].Exporter = func(v any, i int) any {
-			switch v := v.(*Industrial); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[26].Exporter = func(v any, i int) any {
-			switch v := v.(*IndustrialGroup); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[27].Exporter = func(v any, i int) any {
-			switch v := v.(*AccountInfo); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[28].Exporter = func(v any, i int) any {
-			switch v := v.(*Address); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[29].Exporter = func(v any, i int) any {
-			switch v := v.(*AdditionalKey); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[30].Exporter = func(v any, i int) any {
-			switch v := v.(*SignedAddress); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[31].Exporter = func(v any, i int) any {
-			switch v := v.(*SignaturePolicy); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[32].Exporter = func(v any, i int) any {
-			switch v := v.(*AclResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[33].Exporter = func(v any, i int) any {
-			switch v := v.(*Nonce); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[34].Exporter = func(v any, i int) any {
-			switch v := v.(*PendingTx); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[35].Exporter = func(v any, i int) any {
-			switch v := v.(*Pair); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[36].Exporter = func(v any, i int) any {
-			switch v := v.(*CCTransfer); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[37].Exporter = func(v any, i int) any {
-			switch v := v.(*CCTransfers); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_batch_proto_msgTypes[38].Exporter = func(v any, i int) any {
-			switch v := v.(*CCTransferItem); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
