@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.34.2
-// 	protoc        v5.28.2
+// 	protoc-gen-go v1.36.1
+// 	protoc        v5.29.3
 // source: foundation_config.proto
 
 package proto
@@ -24,22 +24,19 @@ const (
 
 // Config stores all chaincode configuration parameters.
 type Config struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Contract      *ContractConfig        `protobuf:"bytes,1,opt,name=contract,proto3" json:"contract,omitempty"`
+	Token         *TokenConfig           `protobuf:"bytes,2,opt,name=token,proto3" json:"token,omitempty"`
+	ExtConfig     *anypb.Any             `protobuf:"bytes,3,opt,name=ext_config,json=extConfig,proto3" json:"ext_config,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Contract  *ContractConfig `protobuf:"bytes,1,opt,name=contract,proto3" json:"contract,omitempty"`
-	Token     *TokenConfig    `protobuf:"bytes,2,opt,name=token,proto3" json:"token,omitempty"`
-	ExtConfig *anypb.Any      `protobuf:"bytes,3,opt,name=ext_config,json=extConfig,proto3" json:"ext_config,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Config) Reset() {
 	*x = Config{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_foundation_config_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_foundation_config_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Config) String() string {
@@ -50,7 +47,7 @@ func (*Config) ProtoMessage() {}
 
 func (x *Config) ProtoReflect() protoreflect.Message {
 	mi := &file_foundation_config_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -88,10 +85,7 @@ func (x *Config) GetExtConfig() *anypb.Any {
 
 // ContractConfig is a configuration for platform base contract.
 type ContractConfig struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// symbol is a contract identifier, usually called token ticker.
 	Symbol string `protobuf:"bytes,1,opt,name=symbol,proto3" json:"symbol,omitempty"`
 	// ChaincodeOptions stores contract options.
@@ -105,15 +99,15 @@ type ContractConfig struct {
 	// maxChannelTransferItems maximum count channel transfer items (token amount) for single transfer operation,
 	// if maxChannelTransferItems is 0 will use default value 100
 	MaxChannelTransferItems uint32 `protobuf:"varint,6,opt,name=maxChannelTransferItems,proto3" json:"maxChannelTransferItems,omitempty"`
+	unknownFields           protoimpl.UnknownFields
+	sizeCache               protoimpl.SizeCache
 }
 
 func (x *ContractConfig) Reset() {
 	*x = ContractConfig{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_foundation_config_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_foundation_config_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ContractConfig) String() string {
@@ -124,7 +118,7 @@ func (*ContractConfig) ProtoMessage() {}
 
 func (x *ContractConfig) ProtoReflect() protoreflect.Message {
 	mi := &file_foundation_config_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -182,23 +176,20 @@ func (x *ContractConfig) GetMaxChannelTransferItems() uint32 {
 }
 
 type CollectorEndpoint struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	Endpoint                 string `protobuf:"bytes,1,opt,name=endpoint,proto3" json:"endpoint,omitempty"`
-	AuthorizationHeaderKey   string `protobuf:"bytes,2,opt,name=authorization_header_key,json=authorizationHeaderKey,proto3" json:"authorization_header_key,omitempty"`
-	AuthorizationHeaderValue string `protobuf:"bytes,3,opt,name=authorization_header_value,json=authorizationHeaderValue,proto3" json:"authorization_header_value,omitempty"`
-	TlsCa                    string `protobuf:"bytes,4,opt,name=tls_ca,json=tlsCa,proto3" json:"tls_ca,omitempty"`
+	state                    protoimpl.MessageState `protogen:"open.v1"`
+	Endpoint                 string                 `protobuf:"bytes,1,opt,name=endpoint,proto3" json:"endpoint,omitempty"`
+	AuthorizationHeaderKey   string                 `protobuf:"bytes,2,opt,name=authorization_header_key,json=authorizationHeaderKey,proto3" json:"authorization_header_key,omitempty"`
+	AuthorizationHeaderValue string                 `protobuf:"bytes,3,opt,name=authorization_header_value,json=authorizationHeaderValue,proto3" json:"authorization_header_value,omitempty"`
+	TlsCa                    string                 `protobuf:"bytes,4,opt,name=tls_ca,json=tlsCa,proto3" json:"tls_ca,omitempty"`
+	unknownFields            protoimpl.UnknownFields
+	sizeCache                protoimpl.SizeCache
 }
 
 func (x *CollectorEndpoint) Reset() {
 	*x = CollectorEndpoint{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_foundation_config_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_foundation_config_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CollectorEndpoint) String() string {
@@ -209,7 +200,7 @@ func (*CollectorEndpoint) ProtoMessage() {}
 
 func (x *CollectorEndpoint) ProtoReflect() protoreflect.Message {
 	mi := &file_foundation_config_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -254,10 +245,7 @@ func (x *CollectorEndpoint) GetTlsCa() string {
 
 // ChaincodeOptions stores possible chaincode configuration options.
 type ChaincodeOptions struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// disabled_functions stores list of disabled chaincode methods.
 	// This methods can't be called.
 	DisabledFunctions []string `protobuf:"bytes,1,rep,name=disabled_functions,json=disabledFunctions,proto3" json:"disabled_functions,omitempty"`
@@ -265,15 +253,15 @@ type ChaincodeOptions struct {
 	DisableSwaps bool `protobuf:"varint,2,opt,name=disable_swaps,json=disableSwaps,proto3" json:"disable_swaps,omitempty"`
 	// disable_multi_swaps determines whether multi-swap operations can be performed.
 	DisableMultiSwaps bool `protobuf:"varint,3,opt,name=disable_multi_swaps,json=disableMultiSwaps,proto3" json:"disable_multi_swaps,omitempty"`
+	unknownFields     protoimpl.UnknownFields
+	sizeCache         protoimpl.SizeCache
 }
 
 func (x *ChaincodeOptions) Reset() {
 	*x = ChaincodeOptions{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_foundation_config_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_foundation_config_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ChaincodeOptions) String() string {
@@ -284,7 +272,7 @@ func (*ChaincodeOptions) ProtoMessage() {}
 
 func (x *ChaincodeOptions) ProtoReflect() protoreflect.Message {
 	mi := &file_foundation_config_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -322,21 +310,18 @@ func (x *ChaincodeOptions) GetDisableMultiSwaps() bool {
 
 // Wallet stores user specific data.
 type Wallet struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// address is a user public key in base58check format.
-	Address string `protobuf:"bytes,1,opt,name=address,proto3" json:"address,omitempty"`
+	Address       string `protobuf:"bytes,1,opt,name=address,proto3" json:"address,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Wallet) Reset() {
 	*x = Wallet{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_foundation_config_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_foundation_config_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Wallet) String() string {
@@ -347,7 +332,7 @@ func (*Wallet) ProtoMessage() {}
 
 func (x *Wallet) ProtoReflect() protoreflect.Message {
 	mi := &file_foundation_config_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -371,10 +356,7 @@ func (x *Wallet) GetAddress() string {
 
 // TokenConfig is a configuration for base token.
 type TokenConfig struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// name is a token name.
 	Name string `protobuf:"bytes,1,opt,name=name,proto3" json:"name,omitempty"`
 	// decimals is a token's dimension.
@@ -388,16 +370,16 @@ type TokenConfig struct {
 	// fee_address_setter is the user who can set the set address of the commissions manager.
 	FeeAddressSetter *Wallet `protobuf:"bytes,6,opt,name=fee_address_setter,json=feeAddressSetter,proto3" json:"fee_address_setter,omitempty"`
 	// redeemer is the user who has permission to manage redemption process.
-	Redeemer *Wallet `protobuf:"bytes,7,opt,name=redeemer,proto3" json:"redeemer,omitempty"`
+	Redeemer      *Wallet `protobuf:"bytes,7,opt,name=redeemer,proto3" json:"redeemer,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *TokenConfig) Reset() {
 	*x = TokenConfig{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_foundation_config_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_foundation_config_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TokenConfig) String() string {
@@ -408,7 +390,7 @@ func (*TokenConfig) ProtoMessage() {}
 
 func (x *TokenConfig) ProtoReflect() protoreflect.Message {
 	mi := &file_foundation_config_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -609,80 +591,6 @@ func init() { file_foundation_config_proto_init() }
 func file_foundation_config_proto_init() {
 	if File_foundation_config_proto != nil {
 		return
-	}
-	if !protoimpl.UnsafeEnabled {
-		file_foundation_config_proto_msgTypes[0].Exporter = func(v any, i int) any {
-			switch v := v.(*Config); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_foundation_config_proto_msgTypes[1].Exporter = func(v any, i int) any {
-			switch v := v.(*ContractConfig); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_foundation_config_proto_msgTypes[2].Exporter = func(v any, i int) any {
-			switch v := v.(*CollectorEndpoint); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_foundation_config_proto_msgTypes[3].Exporter = func(v any, i int) any {
-			switch v := v.(*ChaincodeOptions); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_foundation_config_proto_msgTypes[4].Exporter = func(v any, i int) any {
-			switch v := v.(*Wallet); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_foundation_config_proto_msgTypes[5].Exporter = func(v any, i int) any {
-			switch v := v.(*TokenConfig); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
