@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.34.2
-// 	protoc        v5.28.2
+// 	protoc-gen-go v1.36.1
+// 	protoc        v5.29.3
 // source: locks.proto
 
 package proto
@@ -21,26 +21,23 @@ const (
 )
 
 type BalanceLockRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Id            string                 `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`           // lock identifier (optional parameter, if not specified - TXID is used)
+	Address       string                 `protobuf:"bytes,2,opt,name=address,proto3" json:"address,omitempty"` // The owner of the owner
+	Token         string                 `protobuf:"bytes,3,opt,name=token,proto3" json:"token,omitempty"`     // id and ticker token
+	Amount        string                 `protobuf:"bytes,4,opt,name=amount,proto3" json:"amount,omitempty"`   // big.Int number of locking tokens
+	Reason        string                 `protobuf:"bytes,5,opt,name=reason,proto3" json:"reason,omitempty"`   // The reason for the blocking
+	Docs          [][]byte               `protobuf:"bytes,6,rep,name=docs,proto3" json:"docs,omitempty"`       // Hash of documents with justification (optional parameter)
+	Payload       []byte                 `protobuf:"bytes,7,opt,name=payload,proto3" json:"payload,omitempty"` // Additional information (optional parameter)
 	unknownFields protoimpl.UnknownFields
-
-	Id      string   `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`           // lock identifier (optional parameter, if not specified - TXID is used)
-	Address string   `protobuf:"bytes,2,opt,name=address,proto3" json:"address,omitempty"` // The owner of the owner
-	Token   string   `protobuf:"bytes,3,opt,name=token,proto3" json:"token,omitempty"`     // id and ticker token
-	Amount  string   `protobuf:"bytes,4,opt,name=amount,proto3" json:"amount,omitempty"`   // big.Int number of locking tokens
-	Reason  string   `protobuf:"bytes,5,opt,name=reason,proto3" json:"reason,omitempty"`   // The reason for the blocking
-	Docs    [][]byte `protobuf:"bytes,6,rep,name=docs,proto3" json:"docs,omitempty"`       // Hash of documents with justification (optional parameter)
-	Payload []byte   `protobuf:"bytes,7,opt,name=payload,proto3" json:"payload,omitempty"` // Additional information (optional parameter)
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *BalanceLockRequest) Reset() {
 	*x = BalanceLockRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_locks_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_locks_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *BalanceLockRequest) String() string {
@@ -51,7 +48,7 @@ func (*BalanceLockRequest) ProtoMessage() {}
 
 func (x *BalanceLockRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_locks_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -117,27 +114,24 @@ func (x *BalanceLockRequest) GetPayload() []byte {
 
 // State: data on blocking the balance sheet
 type TokenBalanceLock struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Id            string                 `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`                                            // lock identifier (optional parameter, if not specified, is used txID)
+	Address       string                 `protobuf:"bytes,2,opt,name=address,proto3" json:"address,omitempty"`                                  // The addres of the owner
+	Token         string                 `protobuf:"bytes,3,opt,name=token,proto3" json:"token,omitempty"`                                      // id and ticker token
+	InitAmount    string                 `protobuf:"bytes,4,opt,name=init_amount,json=initAmount,proto3" json:"init_amount,omitempty"`          // big.Int the initial number of tokens for locking
+	CurrentAmount string                 `protobuf:"bytes,5,opt,name=current_amount,json=currentAmount,proto3" json:"current_amount,omitempty"` // big.Int The current amount of blocked tokens
+	Reason        string                 `protobuf:"bytes,6,opt,name=reason,proto3" json:"reason,omitempty"`                                    // The reason for the blocking
+	Docs          [][]byte               `protobuf:"bytes,7,rep,name=docs,proto3" json:"docs,omitempty"`                                        // hashes justification documents (optional parameter)
+	Payload       []byte                 `protobuf:"bytes,8,opt,name=payload,proto3" json:"payload,omitempty"`                                  // Additional information (optional parameter)
 	unknownFields protoimpl.UnknownFields
-
-	Id            string   `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`                                            // lock identifier (optional parameter, if not specified, is used txID)
-	Address       string   `protobuf:"bytes,2,opt,name=address,proto3" json:"address,omitempty"`                                  // The addres of the owner
-	Token         string   `protobuf:"bytes,3,opt,name=token,proto3" json:"token,omitempty"`                                      // id and ticker token
-	InitAmount    string   `protobuf:"bytes,4,opt,name=init_amount,json=initAmount,proto3" json:"init_amount,omitempty"`          // big.Int the initial number of tokens for locking
-	CurrentAmount string   `protobuf:"bytes,5,opt,name=current_amount,json=currentAmount,proto3" json:"current_amount,omitempty"` // big.Int The current amount of blocked tokens
-	Reason        string   `protobuf:"bytes,6,opt,name=reason,proto3" json:"reason,omitempty"`                                    // The reason for the blocking
-	Docs          [][]byte `protobuf:"bytes,7,rep,name=docs,proto3" json:"docs,omitempty"`                                        // hashes justification documents (optional parameter)
-	Payload       []byte   `protobuf:"bytes,8,opt,name=payload,proto3" json:"payload,omitempty"`                                  // Additional information (optional parameter)
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *TokenBalanceLock) Reset() {
 	*x = TokenBalanceLock{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_locks_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_locks_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TokenBalanceLock) String() string {
@@ -148,7 +142,7 @@ func (*TokenBalanceLock) ProtoMessage() {}
 
 func (x *TokenBalanceLock) ProtoReflect() protoreflect.Message {
 	mi := &file_locks_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -221,27 +215,24 @@ func (x *TokenBalanceLock) GetPayload() []byte {
 
 // State: data on blocking alved balance
 type AllowedBalanceLock struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Id            string                 `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`                                            // lock identifier (optional parameter, if not specified, is used txID)
+	Address       string                 `protobuf:"bytes,2,opt,name=address,proto3" json:"address,omitempty"`                                  // owner address
+	Token         string                 `protobuf:"bytes,3,opt,name=token,proto3" json:"token,omitempty"`                                      // identifier / tiker token
+	InitAmount    string                 `protobuf:"bytes,4,opt,name=init_amount,json=initAmount,proto3" json:"init_amount,omitempty"`          // big.Int the initial number of tokens for locking
+	CurrentAmount string                 `protobuf:"bytes,5,opt,name=current_amount,json=currentAmount,proto3" json:"current_amount,omitempty"` // big.Int the current amount of blocked tokens
+	Reason        string                 `protobuf:"bytes,6,opt,name=reason,proto3" json:"reason,omitempty"`                                    // the reason for the blocking
+	Docs          [][]byte               `protobuf:"bytes,7,rep,name=docs,proto3" json:"docs,omitempty"`                                        // hashes justification documents (optional parameter)
+	Payload       []byte                 `protobuf:"bytes,8,opt,name=payload,proto3" json:"payload,omitempty"`                                  // Additional information (optional parameter)
 	unknownFields protoimpl.UnknownFields
-
-	Id            string   `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`                                            // lock identifier (optional parameter, if not specified, is used txID)
-	Address       string   `protobuf:"bytes,2,opt,name=address,proto3" json:"address,omitempty"`                                  // owner address
-	Token         string   `protobuf:"bytes,3,opt,name=token,proto3" json:"token,omitempty"`                                      // identifier / tiker token
-	InitAmount    string   `protobuf:"bytes,4,opt,name=init_amount,json=initAmount,proto3" json:"init_amount,omitempty"`          // big.Int the initial number of tokens for locking
-	CurrentAmount string   `protobuf:"bytes,5,opt,name=current_amount,json=currentAmount,proto3" json:"current_amount,omitempty"` // big.Int the current amount of blocked tokens
-	Reason        string   `protobuf:"bytes,6,opt,name=reason,proto3" json:"reason,omitempty"`                                    // the reason for the blocking
-	Docs          [][]byte `protobuf:"bytes,7,rep,name=docs,proto3" json:"docs,omitempty"`                                        // hashes justification documents (optional parameter)
-	Payload       []byte   `protobuf:"bytes,8,opt,name=payload,proto3" json:"payload,omitempty"`                                  // Additional information (optional parameter)
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *AllowedBalanceLock) Reset() {
 	*x = AllowedBalanceLock{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_locks_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_locks_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *AllowedBalanceLock) String() string {
@@ -252,7 +243,7 @@ func (*AllowedBalanceLock) ProtoMessage() {}
 
 func (x *AllowedBalanceLock) ProtoReflect() protoreflect.Message {
 	mi := &file_locks_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -325,26 +316,23 @@ func (x *AllowedBalanceLock) GetPayload() []byte {
 
 // Event: Token balance is blocked
 type TokenBalanceLocked struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Id            string                 `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`           // Blocking identifier
+	Address       string                 `protobuf:"bytes,2,opt,name=address,proto3" json:"address,omitempty"` // owner address
+	Token         string                 `protobuf:"bytes,3,opt,name=token,proto3" json:"token,omitempty"`     // identifier / token tiker
+	Amount        string                 `protobuf:"bytes,4,opt,name=amount,proto3" json:"amount,omitempty"`   // big.Int number of tokens to lock
+	Reason        string                 `protobuf:"bytes,5,opt,name=reason,proto3" json:"reason,omitempty"`   // The reason of blocking
+	Docs          [][]byte               `protobuf:"bytes,6,rep,name=docs,proto3" json:"docs,omitempty"`       // hashes of documents with justification (optional parameter)
+	Payload       []byte                 `protobuf:"bytes,7,opt,name=payload,proto3" json:"payload,omitempty"` // additional information (optional parameter)
 	unknownFields protoimpl.UnknownFields
-
-	Id      string   `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`           // Blocking identifier
-	Address string   `protobuf:"bytes,2,opt,name=address,proto3" json:"address,omitempty"` // owner address
-	Token   string   `protobuf:"bytes,3,opt,name=token,proto3" json:"token,omitempty"`     // identifier / token tiker
-	Amount  string   `protobuf:"bytes,4,opt,name=amount,proto3" json:"amount,omitempty"`   // big.Int number of tokens to lock
-	Reason  string   `protobuf:"bytes,5,opt,name=reason,proto3" json:"reason,omitempty"`   // The reason of blocking
-	Docs    [][]byte `protobuf:"bytes,6,rep,name=docs,proto3" json:"docs,omitempty"`       // hashes of documents with justification (optional parameter)
-	Payload []byte   `protobuf:"bytes,7,opt,name=payload,proto3" json:"payload,omitempty"` // additional information (optional parameter)
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *TokenBalanceLocked) Reset() {
 	*x = TokenBalanceLocked{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_locks_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_locks_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TokenBalanceLocked) String() string {
@@ -355,7 +343,7 @@ func (*TokenBalanceLocked) ProtoMessage() {}
 
 func (x *TokenBalanceLocked) ProtoReflect() protoreflect.Message {
 	mi := &file_locks_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -421,27 +409,24 @@ func (x *TokenBalanceLocked) GetPayload() []byte {
 
 // Event: token balance unlocked
 type TokenBalanceUnlocked struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	Id                string   `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`                                                         // lock identifier
-	Address           string   `protobuf:"bytes,2,opt,name=address,proto3" json:"address,omitempty"`                                               // owner's address
-	Token             string   `protobuf:"bytes,3,opt,name=token,proto3" json:"token,omitempty"`                                                   // token identifier / token ticker
-	Amount            string   `protobuf:"bytes,4,opt,name=amount,proto3" json:"amount,omitempty"`                                                 // big.Int number of tokens to unlock
-	Reason            string   `protobuf:"bytes,5,opt,name=reason,proto3" json:"reason,omitempty"`                                                 // blocking reason
-	Docs              [][]byte `protobuf:"bytes,6,rep,name=docs,proto3" json:"docs,omitempty"`                                                     // hashes of documents with justification (optional parameter)
-	Payload           []byte   `protobuf:"bytes,7,opt,name=payload,proto3" json:"payload,omitempty"`                                               // additional information (optional parameter)
-	CompleteOperation bool     `protobuf:"varint,8,opt,name=complete_operation,json=completeOperation,proto3" json:"complete_operation,omitempty"` // a indication that it's fully unlocked
+	state             protoimpl.MessageState `protogen:"open.v1"`
+	Id                string                 `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`                                                         // lock identifier
+	Address           string                 `protobuf:"bytes,2,opt,name=address,proto3" json:"address,omitempty"`                                               // owner's address
+	Token             string                 `protobuf:"bytes,3,opt,name=token,proto3" json:"token,omitempty"`                                                   // token identifier / token ticker
+	Amount            string                 `protobuf:"bytes,4,opt,name=amount,proto3" json:"amount,omitempty"`                                                 // big.Int number of tokens to unlock
+	Reason            string                 `protobuf:"bytes,5,opt,name=reason,proto3" json:"reason,omitempty"`                                                 // blocking reason
+	Docs              [][]byte               `protobuf:"bytes,6,rep,name=docs,proto3" json:"docs,omitempty"`                                                     // hashes of documents with justification (optional parameter)
+	Payload           []byte                 `protobuf:"bytes,7,opt,name=payload,proto3" json:"payload,omitempty"`                                               // additional information (optional parameter)
+	CompleteOperation bool                   `protobuf:"varint,8,opt,name=complete_operation,json=completeOperation,proto3" json:"complete_operation,omitempty"` // a indication that it's fully unlocked
+	unknownFields     protoimpl.UnknownFields
+	sizeCache         protoimpl.SizeCache
 }
 
 func (x *TokenBalanceUnlocked) Reset() {
 	*x = TokenBalanceUnlocked{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_locks_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_locks_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TokenBalanceUnlocked) String() string {
@@ -452,7 +437,7 @@ func (*TokenBalanceUnlocked) ProtoMessage() {}
 
 func (x *TokenBalanceUnlocked) ProtoReflect() protoreflect.Message {
 	mi := &file_locks_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -525,26 +510,23 @@ func (x *TokenBalanceUnlocked) GetCompleteOperation() bool {
 
 // Event: token balance blocked
 type AllowedBalanceLocked struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Id            string                 `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`           // lock identifier
+	Address       string                 `protobuf:"bytes,2,opt,name=address,proto3" json:"address,omitempty"` // owner's address
+	Token         string                 `protobuf:"bytes,3,opt,name=token,proto3" json:"token,omitempty"`     // token identifier / token ticker
+	Amount        string                 `protobuf:"bytes,4,opt,name=amount,proto3" json:"amount,omitempty"`   // big.Int number of tokens to lock
+	Reason        string                 `protobuf:"bytes,5,opt,name=reason,proto3" json:"reason,omitempty"`   // blocking reason
+	Docs          [][]byte               `protobuf:"bytes,6,rep,name=docs,proto3" json:"docs,omitempty"`       // hashes of documents with justification (optional parameter)
+	Payload       []byte                 `protobuf:"bytes,7,opt,name=payload,proto3" json:"payload,omitempty"` // additional information (optional parameter)
 	unknownFields protoimpl.UnknownFields
-
-	Id      string   `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`           // lock identifier
-	Address string   `protobuf:"bytes,2,opt,name=address,proto3" json:"address,omitempty"` // owner's address
-	Token   string   `protobuf:"bytes,3,opt,name=token,proto3" json:"token,omitempty"`     // token identifier / token ticker
-	Amount  string   `protobuf:"bytes,4,opt,name=amount,proto3" json:"amount,omitempty"`   // big.Int number of tokens to lock
-	Reason  string   `protobuf:"bytes,5,opt,name=reason,proto3" json:"reason,omitempty"`   // blocking reason
-	Docs    [][]byte `protobuf:"bytes,6,rep,name=docs,proto3" json:"docs,omitempty"`       // hashes of documents with justification (optional parameter)
-	Payload []byte   `protobuf:"bytes,7,opt,name=payload,proto3" json:"payload,omitempty"` // additional information (optional parameter)
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *AllowedBalanceLocked) Reset() {
 	*x = AllowedBalanceLocked{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_locks_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_locks_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *AllowedBalanceLocked) String() string {
@@ -555,7 +537,7 @@ func (*AllowedBalanceLocked) ProtoMessage() {}
 
 func (x *AllowedBalanceLocked) ProtoReflect() protoreflect.Message {
 	mi := &file_locks_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -621,27 +603,24 @@ func (x *AllowedBalanceLocked) GetPayload() []byte {
 
 // Event: token balance unlocked
 type AllowedBalanceUnlocked struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	Id                string   `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`                                                         // lock identifier
-	Address           string   `protobuf:"bytes,2,opt,name=address,proto3" json:"address,omitempty"`                                               // owner's address
-	Token             string   `protobuf:"bytes,3,opt,name=token,proto3" json:"token,omitempty"`                                                   // token identifier / token ticker
-	Amount            string   `protobuf:"bytes,4,opt,name=amount,proto3" json:"amount,omitempty"`                                                 // big.Int number of tokens to lock
-	Reason            string   `protobuf:"bytes,5,opt,name=reason,proto3" json:"reason,omitempty"`                                                 // blocking reason
-	Docs              [][]byte `protobuf:"bytes,6,rep,name=docs,proto3" json:"docs,omitempty"`                                                     // hashes of documents with justification (optional parameter)
-	Payload           []byte   `protobuf:"bytes,7,opt,name=payload,proto3" json:"payload,omitempty"`                                               // additional information (optional parameter)
-	CompleteOperation bool     `protobuf:"varint,8,opt,name=complete_operation,json=completeOperation,proto3" json:"complete_operation,omitempty"` // a sign that it's completely unlocked
+	state             protoimpl.MessageState `protogen:"open.v1"`
+	Id                string                 `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`                                                         // lock identifier
+	Address           string                 `protobuf:"bytes,2,opt,name=address,proto3" json:"address,omitempty"`                                               // owner's address
+	Token             string                 `protobuf:"bytes,3,opt,name=token,proto3" json:"token,omitempty"`                                                   // token identifier / token ticker
+	Amount            string                 `protobuf:"bytes,4,opt,name=amount,proto3" json:"amount,omitempty"`                                                 // big.Int number of tokens to lock
+	Reason            string                 `protobuf:"bytes,5,opt,name=reason,proto3" json:"reason,omitempty"`                                                 // blocking reason
+	Docs              [][]byte               `protobuf:"bytes,6,rep,name=docs,proto3" json:"docs,omitempty"`                                                     // hashes of documents with justification (optional parameter)
+	Payload           []byte                 `protobuf:"bytes,7,opt,name=payload,proto3" json:"payload,omitempty"`                                               // additional information (optional parameter)
+	CompleteOperation bool                   `protobuf:"varint,8,opt,name=complete_operation,json=completeOperation,proto3" json:"complete_operation,omitempty"` // a sign that it's completely unlocked
+	unknownFields     protoimpl.UnknownFields
+	sizeCache         protoimpl.SizeCache
 }
 
 func (x *AllowedBalanceUnlocked) Reset() {
 	*x = AllowedBalanceUnlocked{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_locks_proto_msgTypes[6]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_locks_proto_msgTypes[6]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *AllowedBalanceUnlocked) String() string {
@@ -652,7 +631,7 @@ func (*AllowedBalanceUnlocked) ProtoMessage() {}
 
 func (x *AllowedBalanceUnlocked) ProtoReflect() protoreflect.Message {
 	mi := &file_locks_proto_msgTypes[6]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -858,92 +837,6 @@ func init() { file_locks_proto_init() }
 func file_locks_proto_init() {
 	if File_locks_proto != nil {
 		return
-	}
-	if !protoimpl.UnsafeEnabled {
-		file_locks_proto_msgTypes[0].Exporter = func(v any, i int) any {
-			switch v := v.(*BalanceLockRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_locks_proto_msgTypes[1].Exporter = func(v any, i int) any {
-			switch v := v.(*TokenBalanceLock); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_locks_proto_msgTypes[2].Exporter = func(v any, i int) any {
-			switch v := v.(*AllowedBalanceLock); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_locks_proto_msgTypes[3].Exporter = func(v any, i int) any {
-			switch v := v.(*TokenBalanceLocked); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_locks_proto_msgTypes[4].Exporter = func(v any, i int) any {
-			switch v := v.(*TokenBalanceUnlocked); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_locks_proto_msgTypes[5].Exporter = func(v any, i int) any {
-			switch v := v.(*AllowedBalanceLocked); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_locks_proto_msgTypes[6].Exporter = func(v any, i int) any {
-			switch v := v.(*AllowedBalanceUnlocked); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
