@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.34.2
-// 	protoc        v5.28.2
+// 	protoc-gen-go v1.36.1
+// 	protoc        v5.29.3
 // source: report.proto
 
 package proto
@@ -21,23 +21,20 @@ const (
 )
 
 type Report struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	FeePublicKey     []byte  `protobuf:"bytes,1,opt,name=fee_public_key,json=feePublicKey,proto3" json:"fee_public_key,omitempty"`
-	ChecksumOrderer  uint64  `protobuf:"varint,2,opt,name=checksum_orderer,json=checksumOrderer,proto3" json:"checksum_orderer,omitempty"`
-	ChecksumEndorser uint64  `protobuf:"varint,3,opt,name=checksum_endorser,json=checksumEndorser,proto3" json:"checksum_endorser,omitempty"`
-	Stats            []*Stat `protobuf:"bytes,4,rep,name=stats,proto3" json:"stats,omitempty"`
+	state            protoimpl.MessageState `protogen:"open.v1"`
+	FeePublicKey     []byte                 `protobuf:"bytes,1,opt,name=fee_public_key,json=feePublicKey,proto3" json:"fee_public_key,omitempty"`
+	ChecksumOrderer  uint64                 `protobuf:"varint,2,opt,name=checksum_orderer,json=checksumOrderer,proto3" json:"checksum_orderer,omitempty"`
+	ChecksumEndorser uint64                 `protobuf:"varint,3,opt,name=checksum_endorser,json=checksumEndorser,proto3" json:"checksum_endorser,omitempty"`
+	Stats            []*Stat                `protobuf:"bytes,4,rep,name=stats,proto3" json:"stats,omitempty"`
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
 }
 
 func (x *Report) Reset() {
 	*x = Report{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_report_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_report_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Report) String() string {
@@ -48,7 +45,7 @@ func (*Report) ProtoMessage() {}
 
 func (x *Report) ProtoReflect() protoreflect.Message {
 	mi := &file_report_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -92,22 +89,19 @@ func (x *Report) GetStats() []*Stat {
 }
 
 type Stat struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	CertOwner      string `protobuf:"bytes,1,opt,name=cert_owner,json=certOwner,proto3" json:"cert_owner,omitempty"`
-	PointsEndorser uint64 `protobuf:"varint,2,opt,name=points_endorser,json=pointsEndorser,proto3" json:"points_endorser,omitempty"`
-	PointsOrderer  uint64 `protobuf:"varint,3,opt,name=points_orderer,json=pointsOrderer,proto3" json:"points_orderer,omitempty"`
+	state          protoimpl.MessageState `protogen:"open.v1"`
+	CertOwner      string                 `protobuf:"bytes,1,opt,name=cert_owner,json=certOwner,proto3" json:"cert_owner,omitempty"`
+	PointsEndorser uint64                 `protobuf:"varint,2,opt,name=points_endorser,json=pointsEndorser,proto3" json:"points_endorser,omitempty"`
+	PointsOrderer  uint64                 `protobuf:"varint,3,opt,name=points_orderer,json=pointsOrderer,proto3" json:"points_orderer,omitempty"`
+	unknownFields  protoimpl.UnknownFields
+	sizeCache      protoimpl.SizeCache
 }
 
 func (x *Stat) Reset() {
 	*x = Stat{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_report_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_report_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Stat) String() string {
@@ -118,7 +112,7 @@ func (*Stat) ProtoMessage() {}
 
 func (x *Stat) ProtoReflect() protoreflect.Message {
 	mi := &file_report_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -155,20 +149,17 @@ func (x *Stat) GetPointsOrderer() uint64 {
 }
 
 type HeadInfo struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Heads         []*Head                `protobuf:"bytes,1,rep,name=heads,proto3" json:"heads,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Heads []*Head `protobuf:"bytes,1,rep,name=heads,proto3" json:"heads,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *HeadInfo) Reset() {
 	*x = HeadInfo{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_report_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_report_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *HeadInfo) String() string {
@@ -179,7 +170,7 @@ func (*HeadInfo) ProtoMessage() {}
 
 func (x *HeadInfo) ProtoReflect() protoreflect.Message {
 	mi := &file_report_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -202,21 +193,18 @@ func (x *HeadInfo) GetHeads() []*Head {
 }
 
 type Head struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Token         string                 `protobuf:"bytes,1,opt,name=token,proto3" json:"token,omitempty"`
+	BlockNum      uint64                 `protobuf:"varint,2,opt,name=block_num,json=blockNum,proto3" json:"block_num,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Token    string `protobuf:"bytes,1,opt,name=token,proto3" json:"token,omitempty"`
-	BlockNum uint64 `protobuf:"varint,2,opt,name=block_num,json=blockNum,proto3" json:"block_num,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Head) Reset() {
 	*x = Head{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_report_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_report_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Head) String() string {
@@ -227,7 +215,7 @@ func (*Head) ProtoMessage() {}
 
 func (x *Head) ProtoReflect() protoreflect.Message {
 	mi := &file_report_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -324,56 +312,6 @@ func init() { file_report_proto_init() }
 func file_report_proto_init() {
 	if File_report_proto != nil {
 		return
-	}
-	if !protoimpl.UnsafeEnabled {
-		file_report_proto_msgTypes[0].Exporter = func(v any, i int) any {
-			switch v := v.(*Report); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_report_proto_msgTypes[1].Exporter = func(v any, i int) any {
-			switch v := v.(*Stat); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_report_proto_msgTypes[2].Exporter = func(v any, i int) any {
-			switch v := v.(*HeadInfo); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_report_proto_msgTypes[3].Exporter = func(v any, i int) any {
-			switch v := v.(*Head); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
