@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.34.2
-// 	protoc        v5.28.2
+// 	protoc-gen-go v1.36.1
+// 	protoc        v5.29.3
 // source: transfer_request.proto
 
 package proto
@@ -181,10 +181,7 @@ func (BalanceType) EnumDescriptor() ([]byte, []int) {
 
 // Message representing a transfer request.
 type TransferRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// Transfer request ID
 	RequestId string `protobuf:"bytes,1,opt,name=request_id,json=requestId,proto3" json:"request_id,omitempty"`
 	// Basis for the transfer
@@ -213,15 +210,15 @@ type TransferRequest struct {
 	BalanceType BalanceType `protobuf:"varint,13,opt,name=balance_type,json=balanceType,proto3,enum=proto.BalanceType" json:"balance_type,omitempty"`
 	// Optional additional information
 	AdditionalInfo *anypb.Any `protobuf:"bytes,15,opt,name=additional_info,json=additionalInfo,proto3" json:"additional_info,omitempty"`
+	unknownFields  protoimpl.UnknownFields
+	sizeCache      protoimpl.SizeCache
 }
 
 func (x *TransferRequest) Reset() {
 	*x = TransferRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_transfer_request_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_transfer_request_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *TransferRequest) String() string {
@@ -232,7 +229,7 @@ func (*TransferRequest) ProtoMessage() {}
 
 func (x *TransferRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_transfer_request_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -471,20 +468,6 @@ func init() { file_transfer_request_proto_init() }
 func file_transfer_request_proto_init() {
 	if File_transfer_request_proto != nil {
 		return
-	}
-	if !protoimpl.UnsafeEnabled {
-		file_transfer_request_proto_msgTypes[0].Exporter = func(v any, i int) any {
-			switch v := v.(*TransferRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
