@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.34.2
-// 	protoc        v5.28.2
+// 	protoc-gen-go v1.36.1
+// 	protoc        v5.29.3
 // source: balance_service.proto
 
 package service
@@ -25,20 +25,17 @@ const (
 
 // Address represents a blockchain address.
 type Address struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Base58Check   string                 `protobuf:"bytes,1,opt,name=base58check,proto3" json:"base58check,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Base58Check string `protobuf:"bytes,1,opt,name=base58check,proto3" json:"base58check,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Address) Reset() {
 	*x = Address{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_balance_service_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_balance_service_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Address) String() string {
@@ -49,7 +46,7 @@ func (*Address) ProtoMessage() {}
 
 func (x *Address) ProtoReflect() protoreflect.Message {
 	mi := &file_balance_service_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -73,20 +70,17 @@ func (x *Address) GetBase58Check() string {
 
 // BigInt represents a large integer.
 type BigInt struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Value         string                 `protobuf:"bytes,1,opt,name=value,proto3" json:"value,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Value string `protobuf:"bytes,1,opt,name=value,proto3" json:"value,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *BigInt) Reset() {
 	*x = BigInt{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_balance_service_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_balance_service_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *BigInt) String() string {
@@ -97,7 +91,7 @@ func (*BigInt) ProtoMessage() {}
 
 func (x *BigInt) ProtoReflect() protoreflect.Message {
 	mi := &file_balance_service_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -120,22 +114,19 @@ func (x *BigInt) GetValue() string {
 }
 
 type BalanceAdjustmentRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Address       *Address               `protobuf:"bytes,1,opt,name=address,proto3" json:"address,omitempty"`
+	Amount        *BigInt                `protobuf:"bytes,2,opt,name=amount,proto3" json:"amount,omitempty"`
+	Reason        string                 `protobuf:"bytes,3,opt,name=reason,proto3" json:"reason,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Address *Address `protobuf:"bytes,1,opt,name=address,proto3" json:"address,omitempty"`
-	Amount  *BigInt  `protobuf:"bytes,2,opt,name=amount,proto3" json:"amount,omitempty"`
-	Reason  string   `protobuf:"bytes,3,opt,name=reason,proto3" json:"reason,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *BalanceAdjustmentRequest) Reset() {
 	*x = BalanceAdjustmentRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_balance_service_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_balance_service_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *BalanceAdjustmentRequest) String() string {
@@ -146,7 +137,7 @@ func (*BalanceAdjustmentRequest) ProtoMessage() {}
 
 func (x *BalanceAdjustmentRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_balance_service_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -261,44 +252,6 @@ func init() { file_balance_service_proto_init() }
 func file_balance_service_proto_init() {
 	if File_balance_service_proto != nil {
 		return
-	}
-	if !protoimpl.UnsafeEnabled {
-		file_balance_service_proto_msgTypes[0].Exporter = func(v any, i int) any {
-			switch v := v.(*Address); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_balance_service_proto_msgTypes[1].Exporter = func(v any, i int) any {
-			switch v := v.(*BigInt); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_balance_service_proto_msgTypes[2].Exporter = func(v any, i int) any {
-			switch v := v.(*BalanceAdjustmentRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
