@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.34.2
-// 	protoc        v5.28.2
+// 	protoc-gen-go v1.36.1
+// 	protoc        v5.29.3
 // source: ext_config.proto
 
 package industrialtoken
@@ -23,30 +23,27 @@ const (
 )
 
 type ExtConfig struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	Name             string        `protobuf:"bytes,1,opt,name=name,proto3" json:"name,omitempty"`
-	Decimals         uint32        `protobuf:"varint,2,opt,name=decimals,proto3" json:"decimals,omitempty"`
-	UnderlyingAsset  string        `protobuf:"bytes,3,opt,name=underlying_asset,json=underlyingAsset,proto3" json:"underlying_asset,omitempty"`
-	DeliveryForm     string        `protobuf:"bytes,4,opt,name=delivery_form,json=deliveryForm,proto3" json:"delivery_form,omitempty"`
-	UnitOfMeasure    string        `protobuf:"bytes,5,opt,name=unit_of_measure,json=unitOfMeasure,proto3" json:"unit_of_measure,omitempty"`
-	TokensForUnit    string        `protobuf:"bytes,6,opt,name=tokens_for_unit,json=tokensForUnit,proto3" json:"tokens_for_unit,omitempty"`
-	PaymentTerms     string        `protobuf:"bytes,7,opt,name=payment_terms,json=paymentTerms,proto3" json:"payment_terms,omitempty"`
-	Price            string        `protobuf:"bytes,8,opt,name=price,proto3" json:"price,omitempty"`
-	Issuer           *proto.Wallet `protobuf:"bytes,9,opt,name=issuer,proto3" json:"issuer,omitempty"`
-	FeeSetter        *proto.Wallet `protobuf:"bytes,10,opt,name=fee_setter,json=feeSetter,proto3" json:"fee_setter,omitempty"`
-	FeeAddressSetter *proto.Wallet `protobuf:"bytes,11,opt,name=fee_address_setter,json=feeAddressSetter,proto3" json:"fee_address_setter,omitempty"`
+	state            protoimpl.MessageState `protogen:"open.v1"`
+	Name             string                 `protobuf:"bytes,1,opt,name=name,proto3" json:"name,omitempty"`
+	Decimals         uint32                 `protobuf:"varint,2,opt,name=decimals,proto3" json:"decimals,omitempty"`
+	UnderlyingAsset  string                 `protobuf:"bytes,3,opt,name=underlying_asset,json=underlyingAsset,proto3" json:"underlying_asset,omitempty"`
+	DeliveryForm     string                 `protobuf:"bytes,4,opt,name=delivery_form,json=deliveryForm,proto3" json:"delivery_form,omitempty"`
+	UnitOfMeasure    string                 `protobuf:"bytes,5,opt,name=unit_of_measure,json=unitOfMeasure,proto3" json:"unit_of_measure,omitempty"`
+	TokensForUnit    string                 `protobuf:"bytes,6,opt,name=tokens_for_unit,json=tokensForUnit,proto3" json:"tokens_for_unit,omitempty"`
+	PaymentTerms     string                 `protobuf:"bytes,7,opt,name=payment_terms,json=paymentTerms,proto3" json:"payment_terms,omitempty"`
+	Price            string                 `protobuf:"bytes,8,opt,name=price,proto3" json:"price,omitempty"`
+	Issuer           *proto.Wallet          `protobuf:"bytes,9,opt,name=issuer,proto3" json:"issuer,omitempty"`
+	FeeSetter        *proto.Wallet          `protobuf:"bytes,10,opt,name=fee_setter,json=feeSetter,proto3" json:"fee_setter,omitempty"`
+	FeeAddressSetter *proto.Wallet          `protobuf:"bytes,11,opt,name=fee_address_setter,json=feeAddressSetter,proto3" json:"fee_address_setter,omitempty"`
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
 }
 
 func (x *ExtConfig) Reset() {
 	*x = ExtConfig{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_ext_config_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_ext_config_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ExtConfig) String() string {
@@ -57,7 +54,7 @@ func (*ExtConfig) ProtoMessage() {}
 
 func (x *ExtConfig) ProtoReflect() protoreflect.Message {
 	mi := &file_ext_config_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -227,20 +224,6 @@ func init() { file_ext_config_proto_init() }
 func file_ext_config_proto_init() {
 	if File_ext_config_proto != nil {
 		return
-	}
-	if !protoimpl.UnsafeEnabled {
-		file_ext_config_proto_msgTypes[0].Exporter = func(v any, i int) any {
-			switch v := v.(*ExtConfig); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
