@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.2.0
-// - protoc             v5.27.1
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.29.3
 // source: balance_service.proto
 
 package proto
@@ -16,12 +16,19 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
+
+const (
+	BalanceService_AddBalanceByAdmin_FullMethodName = "/foundation.token.BalanceService/AddBalanceByAdmin"
+	BalanceService_HelloWorld_FullMethodName        = "/foundation.token.BalanceService/HelloWorld"
+)
 
 // BalanceServiceClient is the client API for BalanceService service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// BalanceService defines the balance service.
 type BalanceServiceClient interface {
 	AddBalanceByAdmin(ctx context.Context, in *BalanceAdjustmentRequest, opts ...grpc.CallOption) (*emptypb.Empty, error)
 	HelloWorld(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*HelloWorldResponse, error)
@@ -36,8 +43,9 @@ func NewBalanceServiceClient(cc grpc.ClientConnInterface) BalanceServiceClient {
 }
 
 func (c *balanceServiceClient) AddBalanceByAdmin(ctx context.Context, in *BalanceAdjustmentRequest, opts ...grpc.CallOption) (*emptypb.Empty, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(emptypb.Empty)
-	err := c.cc.Invoke(ctx, "/foundation.token.BalanceService/AddBalanceByAdmin", in, out, opts...)
+	err := c.cc.Invoke(ctx, BalanceService_AddBalanceByAdmin_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -45,8 +53,9 @@ func (c *balanceServiceClient) AddBalanceByAdmin(ctx context.Context, in *Balanc
 }
 
 func (c *balanceServiceClient) HelloWorld(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*HelloWorldResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(HelloWorldResponse)
-	err := c.cc.Invoke(ctx, "/foundation.token.BalanceService/HelloWorld", in, out, opts...)
+	err := c.cc.Invoke(ctx, BalanceService_HelloWorld_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -55,16 +64,21 @@ func (c *balanceServiceClient) HelloWorld(ctx context.Context, in *emptypb.Empty
 
 // BalanceServiceServer is the server API for BalanceService service.
 // All implementations must embed UnimplementedBalanceServiceServer
-// for forward compatibility
+// for forward compatibility.
+//
+// BalanceService defines the balance service.
 type BalanceServiceServer interface {
 	AddBalanceByAdmin(context.Context, *BalanceAdjustmentRequest) (*emptypb.Empty, error)
 	HelloWorld(context.Context, *emptypb.Empty) (*HelloWorldResponse, error)
 	mustEmbedUnimplementedBalanceServiceServer()
 }
 
-// UnimplementedBalanceServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedBalanceServiceServer struct {
-}
+// UnimplementedBalanceServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedBalanceServiceServer struct{}
 
 func (UnimplementedBalanceServiceServer) AddBalanceByAdmin(context.Context, *BalanceAdjustmentRequest) (*emptypb.Empty, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method AddBalanceByAdmin not implemented")
@@ -73,6 +87,7 @@ func (UnimplementedBalanceServiceServer) HelloWorld(context.Context, *emptypb.Em
 	return nil, status.Errorf(codes.Unimplemented, "method HelloWorld not implemented")
 }
 func (UnimplementedBalanceServiceServer) mustEmbedUnimplementedBalanceServiceServer() {}
+func (UnimplementedBalanceServiceServer) testEmbeddedByValue()                        {}
 
 // UnsafeBalanceServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to BalanceServiceServer will
@@ -82,6 +97,13 @@ type UnsafeBalanceServiceServer interface {
 }
 
 func RegisterBalanceServiceServer(s grpc.ServiceRegistrar, srv BalanceServiceServer) {
+	// If the following call pancis, it indicates UnimplementedBalanceServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&BalanceService_ServiceDesc, srv)
 }
 
@@ -95,7 +117,7 @@ func _BalanceService_AddBalanceByAdmin_Handler(srv interface{}, ctx context.Cont
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/foundation.token.BalanceService/AddBalanceByAdmin",
+		FullMethod: BalanceService_AddBalanceByAdmin_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(BalanceServiceServer).AddBalanceByAdmin(ctx, req.(*BalanceAdjustmentRequest))
@@ -113,7 +135,7 @@ func _BalanceService_HelloWorld_Handler(srv interface{}, ctx context.Context, de
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/foundation.token.BalanceService/HelloWorld",
+		FullMethod: BalanceService_HelloWorld_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(BalanceServiceServer).HelloWorld(ctx, req.(*emptypb.Empty))
